Improve process handling and Unreal data retrieval
Refactored process reload logic in Main.cs for reliability and updated MainShared.cs to check for exited processes. Enhanced UnrealEngine utilities with FNamePool property and improved data scanning patterns in DataRetriever. Fixed address calculation in TMemory.ScanRel. Updated post-build event to use a batch script.

+++ Uhara/Code/Utils/Memory.cs
@@ -488,7 +488,7 @@ internal class TMemory : MainShared
                 if (searchOffset != -1)
                 {
                     ulong searchAddress = section[0] + (ulong)searchOffset;
-                    ulong relativeAddress = searchAddress + (ulong)offset;
+                    ulong relativeAddress = (ulong)((long)searchAddress + offset);
                     long relativeValue = ReadMemory<int>(ProcessInstance, relativeAddress);
                     ulong destinationAddress = (ulong)((long)searchAddress + relativeValue + offset + 4);
                     return destinationAddress;
